requirements.txt
asgiref==3.7.2
boto3==1.34.41
botocore==1.34.41
certifi==2024.2.2
Django==5.0.2
django-storages==1.14.2
jmespath==1.0.1
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
